# models/user.openapi.json
{
    "openapi": "3.0.2",
    "info": {
      "title": "User API",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "/users"
      }
    ],
    "paths": {
      "/users": {
        "get": {
          "summary": "Get all users",
          "description": "",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {}
              },
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "description": "",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {}
            },
            "schema": {
              "$ref": "#/components/schemas/CreateRequest"
            }
          },
          "responses": {
            "201": {
              "description": "Created",
              "content": {
                "application/json": {}
              },
              "schema": {
                "$ref": "#/components/schemas/CreateUserResponse"
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Get a user by ID",
          "description": "",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {}
              },
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "CreateRequest": {
          "$ref": "#/components/schemas/User"
        },
        "CreateUserResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            }
          }
        }
      }
    }
  }